simple-crud-django-master/requirements.txt
Django==4.2
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
olefile==0.47
Pillow==10.4.0
pytz==2024.1
gunicorn
whitenoise
dj-database-url
python-dotenv
django-cors-headers
djangorestframework-simplejwt
pytest
pytest-django
coverage
django-debug-toolbar
django-extensions
requests